twister: move handlers class out
Split handlers into own file.
