+++ scripts/pylib/twister/hardwaremap.py
@@ -6,7 +6,31 @@
 
 import os
 from multiprocessing import Lock, Value
+import re
 from enviornment import ZEPHYR_BASE
+import platform
+import yaml
+import scl
+import logging
+from pathlib import Path
+import expr_parser
+
+try:
+    # Use the C LibYAML parser if available, rather than the Python parser.
+    # It's much faster.
+    from yaml import CSafeLoader as SafeLoader
+    from yaml import CDumper as Dumper
+except ImportError:
+    from yaml import SafeLoader, Dumper
+
+try:
+    from tabulate import tabulate
+except ImportError:
+    print("Install tabulate python module with pip to use --device-testing option.")
+
+logger = logging.getLogger('twister')
+logger.setLevel(logging.DEBUG)
+
 
 class DUT(object):
     def __init__(self,
